switchng to typescript
