+++ playwright-action/helpers/constants.ts
@@ -1,34 +1,32 @@
-import { id } from "./playwrightUtils.js";
-
-export const DESIGNER = "designer"
-export const VISUALIZER = "visualizer"
+export const DESIGNER = "designer";
+export const VISUALIZER = "visualizer";
 
 //Objects defining paths, aliases, and wait conditions for intercepting certain API requests.
 export const extension = {
-  path: "/api/provider/extension*",
-  alias: "extensionFileLoad"
+	path: "/api/provider/extension*",
+	alias: "extensionFileLoad",
 };
 
 export const designEndpoint = {
-  path: "/api/pattern*",
-  alias: "designEp",
-  wait: "@designEp",
-  absolutePath: "/api/pattern"
-}
+	path: "/api/pattern*",
+	alias: "designEp",
+	wait: "@designEp",
+	absolutePath: "/api/pattern",
+};
 
 export const MESHMAP_PATH = "/extension/meshmap";
 
-export const CANVAS_CONTAINER_ID = "py-canvas-container"
+export const CANVAS_CONTAINER_ID = "py-canvas-container";
 
 export const TIME = {
-  SMALL: 500,
-  MEDIUM: 1000,
-  LARGE: 1500,
-  XLARGE: 2000,
-  XXLARGE: 5000,
-  XXXLARGE: 10000,
-  X4LARGE: 15_000
-}
+	SMALL: 500,
+	MEDIUM: 1000,
+	LARGE: 1500,
+	XLARGE: 2000,
+	XXLARGE: 5000,
+	XXXLARGE: 10000,
+	X4LARGE: 15_000,
+};
 
 // export const canvasContainer = {
 //   query: id(CANVAS_CONTAINER_ID),
@@ -36,9 +34,9 @@ export const TIME = {
 // }
 
 export const playwrightTestDesign = {
-  url: "/extension/meshmap?design=142f0054-d9ae-4352-8618-887104a81928",
-  id: "142f0054-d9ae-4352-8618-887104a81928",
-}
+	url: "/extension/meshmap?design=142f0054-d9ae-4352-8618-887104a81928",
+	id: "142f0054-d9ae-4352-8618-887104a81928",
+};
 //Objects related to specific designs or patterns within the application.
 export const hierarchyRelationshipDesign = playwrightTestDesign;
 
@@ -54,22 +52,22 @@ export const hierarchyRelationshipDesign = playwrightTestDesign;
  * Selection and general Event Binding Layer
  */
 export const canvasLayer0 = {
-  query: '[data-id="layer0-selectbox"]',
-  alias: "layer0"
-}
+	query: '[data-id="layer0-selectbox"]',
+	alias: "layer0",
+};
 
 /**
  * drag and drop Layer
  */
 export const canvasLayer1 = {
-  query: '[data-id="layer1-drag"]',
-  alias: "layer1"
-}
+	query: '[data-id="layer1-drag"]',
+	alias: "layer1",
+};
 
 /**
  * Node and Element Layer
  */
 export const canvaslayer2 = {
-  query: '[data-id="layer2-node"]',
-  alias: "layer2"
-}
+	query: '[data-id="layer2-node"]',
+	alias: "layer2",
+};
